rename LLMStep to LLMStructuredStep
