+++ superpipe/steps/__init__.py
@@ -1,6 +1,6 @@
 from .step import *
 from .embedding_classification import *
-from .llm_step import *
+from .llm_structured import *
 from .custom import *
 from .serp import *
-from .simple_llm_step import *
+from .llm_step import *
